- fixed task controller, updated documentation - added option to daqmx to auto adjust oversampling given a target sampling rate of the DAQ - display of NI vision is done in an external window because including it within the workspace was causing sometimes freezing behaviour.

+++ Framework/HW Triggering/HWTriggering.c
@@ -275,7 +275,7 @@ int WaitForHWTrigArmedSlaves (HWTrigMaster_type* master, char** errorInfo)
 	
 	if (!nSlaves) return 0;
 	
-	fCallResult = WaitForMultipleObjects(nSlaves, master->slaveArmedHndls, TRUE, WaitForHWTrigArmedSlaves_Timeout);
+	fCallResult = WaitForMultipleObjects(nSlaves, master->slaveArmedHndls, TRUE, (DWORD) WaitForHWTrigArmedSlaves_Timeout);
 	
 	// check result
 	switch (fCallResult) {
